Folder structure
- Creating folder structure
- Adding card list styling
- Removing Vite css
