+++ monsters-rolodex/src/components/card-list/card-list.styles.css
@@ -1,0 +1,7 @@
+.card-list-container {
+    width: 85vw;
+    margin: 0 auto;
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr 1fr;
+    grid-gap: 20px;
+}
